+++ CHANGELOG.md
@@ -7,6 +7,7 @@ master
 * Updated and fixed background image for DMG file. Fix #61.
 * Added menu item to search for updates. Fix #54.
 * Log with Growl on 10.7, with notification center on 10.8 and up. Fix #59.
+* Use windowDidLoad rather than override showWindow. Fix #62.
 
 v2.0 (8/22/2014)
 ----------------