+++ CHANGELOG.md
@@ -6,12 +6,12 @@ master
 * New icon, retina and dark mode compatible. Fix #48. Fix #41.
 * Updated and fixed background image for DMG file. Fix #61.
 * Added menu item to search for updates. Fix #54.
+* Check for updates automatically, and provide option to disable it. Fix #63, fix #55.
 * Log with Growl on 10.7, with notification center on 10.8 and up. Fix #59.
 * Use windowDidLoad rather than override showWindow. Fix #62.
 * Never die while processing a screenshot, log exceptions as needed. Fix #38.
 * Parse timestamp out of screenshot filename to not reprocess old files. Fix #40.
 * Reflect name randomization in example URL. Fix #60.
-* Check for updates automatically, and provide option to disable it. Fix #63, fix #55.
 
 v2.0 (8/22/2014)
 ----------------