+++ CHANGELOG.md
@@ -3,6 +3,9 @@ UpShot Changelog
 
 master
 ------
+
+v2.1 (11/7/2014)
+----------------
 * New icon, retina and dark mode compatible. Fix #48. Fix #41.
 * Updated and fixed background image for DMG file. Fix #61.
 * Added menu item to search for updates. Fix #54.